utils can't depend on core it seems; move dct
